Fixing up the AgML API (#14)
* Fix class_to_num mappings for bean_disease_uganda and rangeland_weeds_australia in data_sources.json

* Fix class_to_num mappings for soybean_weed_uav_brazil in data_sources.json

* Fixed sugarcane_damage_usa and crop_weeds_greece

* Fixed plant_seedlings_aarhus

* Fixed fruit_detection_worldwide

* Fixed apple_detection_usa

* Removed nonexistent datasets

* Added segmentation visualization utilities

* Fixed backend issues

* Added and fixed dataloader split functionality

* Added missing n_images

+++ agml/_internal/downloads.py
@@ -31,8 +31,8 @@ def download_dataset(dataset_name, dest_dir):
     except botocore.exceptions.ClientError as ce:
         if "Not Found" in str(ce):
             raise ValueError(
-                f"The dataset '{dataset_name}' could not be found in "
-                f"the bucket, perhaps it has not been uploaded yet.")
+                f"The dataset '{dataset_name}' could not be found in the bucket, "
+                f"perhaps it has not been uploaded yet. Please report this issue.")
         raise ce
 
     # File path of zipped dataset
@@ -60,4 +60,5 @@ def download_dataset(dataset_name, dest_dir):
         print('Done!')
 
     # Delete zipped file
-    os.remove(dataset_download_path)
+    if os.path.exists(dataset_download_path):
+        os.remove(dataset_download_path)
